+++ apps/web/app/(app)/environments/[environmentId]/product/look/components/ThemeStylingPreviewSurvey.tsx
@@ -174,6 +174,7 @@ export const ThemeStylingPreviewSurvey = ({
                 getSetQuestionId={(f: (value: string) => void) => {
                   setQuestionId = f;
                 }}
+                isMultiLanguageAllowed={false}
               />
             </Modal>
           ) : (
@@ -196,6 +197,7 @@ export const ThemeStylingPreviewSurvey = ({
                   getSetQuestionId={(f: (value: string) => void) => {
                     setQuestionId = f;
                   }}
+                  isMultiLanguageAllowed={false}
                 />
               </div>
             </MediaBackground>
