+++ apps/web/app/s/[surveyId]/components/PinScreen.tsx
@@ -26,6 +26,7 @@ interface PinScreenProps {
   IS_FORMBRICKS_CLOUD: boolean;
   verifiedEmail?: string;
   languageCode: string;
+  isMultiLanguageAllowed: boolean;
 }
 
 export const PinScreen = (props: PinScreenProps) => {
@@ -42,6 +43,7 @@ export const PinScreen = (props: PinScreenProps) => {
     IS_FORMBRICKS_CLOUD,
     verifiedEmail,
     languageCode,
+    isMultiLanguageAllowed,
   } = props;
 
   const [localPinEntry, setLocalPinEntry] = useState<string>("");
@@ -122,6 +124,7 @@ export const PinScreen = (props: PinScreenProps) => {
           webAppUrl={webAppUrl}
           verifiedEmail={verifiedEmail}
           languageCode={languageCode}
+          isMultiLanguageAllowed={isMultiLanguageAllowed}
         />
       </MediaBackground>
       <LegalFooter
