+++ apps/web/app/s/[surveyId]/page.tsx
@@ -170,6 +170,7 @@ const Page = async ({ params, searchParams }: LinkSurveyPageProps) => {
         IS_FORMBRICKS_CLOUD={IS_FORMBRICKS_CLOUD}
         verifiedEmail={verifiedEmail}
         languageCode={languageCode}
+        isMultiLanguageAllowed={isMultiLanguageAllowed}
       />
     );
   }
@@ -188,6 +189,7 @@ const Page = async ({ params, searchParams }: LinkSurveyPageProps) => {
           responseCount={survey.welcomeCard.showResponseCount ? responseCount : undefined}
           verifiedEmail={verifiedEmail}
           languageCode={languageCode}
+          isMultiLanguageAllowed={isMultiLanguageAllowed}
         />
         <LegalFooter
           IMPRINT_URL={IMPRINT_URL}
