+++ packages/ee/multiLanguage/components/DefaultLanguageSelect.tsx
@@ -1,8 +1,8 @@
+import { getLanguageLabel } from "@formbricks/lib/i18n/utils";
 import { TLanguage, TProduct } from "@formbricks/types/product";
 import { DefaultTag } from "@formbricks/ui/DefaultTag";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@formbricks/ui/Select";
 
-import { getLanguageLabel } from "../lib/isoLanguages";
 import { ConfirmationModalProps } from "./MultiLanguageCard";
 
 interface DefaultLanguageSelectProps {
