+++ packages/ee/multiLanguage/components/LanguageToggle.tsx
@@ -1,8 +1,7 @@
+import { getLanguageLabel } from "@formbricks/lib/i18n/utils";
 import { TLanguage } from "@formbricks/types/product";
 import { Label } from "@formbricks/ui/Label";
 import { Switch } from "@formbricks/ui/Switch";
-
-import { getLanguageLabel } from "../lib/isoLanguages";
 
 interface LanguageToggleProps {
   language: TLanguage;
