+++ packages/surveys/src/components/general/SurveyModal.tsx
@@ -23,6 +23,7 @@ export const SurveyModal = ({
   languageCode,
   responseCount,
   styling,
+  isMultiLanguageAllowed,
 }: SurveyModalProps) => {
   const [isOpen, setIsOpen] = useState(true);
 
@@ -70,6 +71,7 @@ export const SurveyModal = ({
           styling={styling}
           isCardBorderVisible={!highlightBorderColor}
           clickOutside={placement === "center" ? clickOutside : undefined}
+          isMultiLanguageAllowed={isMultiLanguageAllowed}
         />
       </Modal>
     </div>
