+++ packages/types/formbricksSurveys.ts
@@ -26,6 +26,7 @@ export interface SurveyBaseProps {
   startAtQuestionId?: string;
   clickOutside?: boolean;
   shouldResetQuestionId?: boolean;
+  isMultiLanguageAllowed: boolean;
 }
 
 export interface SurveyInlineProps extends SurveyBaseProps {
