+++ packages/ui/PreviewSurvey/index.tsx
@@ -258,6 +258,7 @@ export const PreviewSurvey = ({
                       setQuestionId = f;
                     }}
                     onFinished={onFinished}
+                    isMultiLanguageAllowed={false}
                   />
                 </Modal>
               ) : (
@@ -278,6 +279,7 @@ export const PreviewSurvey = ({
                       getSetQuestionId={(f: (value: string) => void) => {
                         setQuestionId = f;
                       }}
+                      isMultiLanguageAllowed={false}
                     />
                   </div>
                 </div>
@@ -343,6 +345,7 @@ export const PreviewSurvey = ({
                     setQuestionId = f;
                   }}
                   onFinished={onFinished}
+                  isMultiLanguageAllowed={false}
                 />
               </Modal>
             ) : (
@@ -364,6 +367,7 @@ export const PreviewSurvey = ({
                     getSetQuestionId={(f: (value: string) => void) => {
                       setQuestionId = f;
                     }}
+                    isMultiLanguageAllowed={false}
                   />
                 </div>
               </MediaBackground>
